+++ x/tokenfactory/keeper/before_send.go
@@ -162,7 +162,10 @@ func (k Keeper) callBeforeSendListener(context context.Context, from, to sdk.Acc
 			}
 			em := sdk.NewEventManager()
 
-			childCtx := ctx.WithGasMeter(storetypes.NewGasMeter(types.BeforeSendHookGasLimit))
+			// Check remaining gas in parent context and use the lesser of the fixed limit and remaining gas
+			gasLimit := min(ctx.GasMeter().GasRemaining(), types.BeforeSendHookGasLimit)
+
+			childCtx := ctx.WithGasMeter(storetypes.NewGasMeter(gasLimit))
 			_, err = k.contractKeeper.Sudo(childCtx.WithEventManager(em), cwAddr, msgBz)
 			if err != nil {
 				if strings.Contains(err.Error(), "no such contract") {